20210323:allenrobel: add_vrf_bgp_neighbor() new method, more...
add_vrf_bgp_neighbor() - use this to add a bgp neighbor to a vrf.  Previously, add_bgp_neighbor() added a bgp neighbor to a vrf, or to the default vrf.  This mandated that bgp neighbors in the default vrf had to be added last.  With the addition of add_vrf_bgp_neighbor(), we now add vrf bgp neighbors to a separate list and so we have removed the requirement that vrf and non-vrf bgp neighbors be added in any particular order.

neighbor_update_source - accept a subinterface as a valid update source.

Updated the documentation to describe the public methods available in this class.

Updated the exampe/unit-test script to align with the above changes.