20210328:allenrobel:require sequence for remark, more...
Make it mandatory that instance.sequence be set when instance.remark is set.

This was needed since NX-OS adds a sequence automatically is sequence is not set by the user.   This would result in a 'duplicate sequence' CLI error when the following was issued (since NX-OS auto-assigns sequence 10 to remark in the below example):

remark foo
10 permit ip 1.1.1.1 0.0.0.255 1.2.2.2 0.0.255.255 dscp af31

Also:

1. Add descriptions for add_ace() and add_acl()
2. Add full example under add_acl()
3. Code cleanup
4 Add test for remark in the unit-test script